pyproject.toml
[tool.poetry]
name = "travelmate"
version = "0.1.0"
description = "A travel planning assistant agent"
authors = ["Wassim Chouchenne <wassimchouchenn@gmail.com>"]
readme = "README.md"
packages = [
    {include = "travelmate", from="."}
]
[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
python-dotenv = "^1.0.1"
crewai = "0.102.0"            
crewai_tools = "0.36.0"
pydantic = "^2.10.6"
litellm = "1.61.8"
unstructured = "0.16.21"      
langchain = "0.3.19"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"